requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
flatbuffers==23.3.3
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grist-api==0.1.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.3.0
jax==0.4.8
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
praw==7.7.0
prawcore==2.3.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redditcleaner==1.1.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
strsimpy==0.2.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-macos==2.12.0
termcolor==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0

opencv-python~=4.7.0.72
tensorflow~=2.12.0
ImageHash~=4.3.1
Pillow~=9.5.0
keybert~=0.7.0
pytesseract~=0.3.10
deepdanbooru~=1.0.0
natsort~=8.3.1
selenium~=4.9.0